requirements.txt
opencv-python
loguru
glfw
pyopengl